--- pyrelational/samplers/__init__.py
@@ -1,7 +1,0 @@
-"""Query samplers module."""
-
-from pyrelational.samplers.samplers import (
-    AbstractSampler,
-    DeterministicSampler,
-    ProbabilisticSampler,
-)
