+++ pyrelational/strategies/classification/entropy_classification_strategy.py
@@ -4,8 +4,8 @@ The score is computed between classes in the posterior predictive distribution t
 choose which observations to propose to the oracle.
 """
 
+from pyrelational.batch_mode_samplers import TopKSampler
 from pyrelational.informativeness import Entropy
-from pyrelational.samplers import DeterministicSampler
 from pyrelational.strategies.classification.classification_strategy import (
     ClassificationStrategy,
 )
@@ -16,4 +16,4 @@ class EntropyClassificationStrategy(ClassificationStrategy):
 
     def __init__(self, axis: int = -1):
         """Initialise the strategy with entropy scorer and deterministic sampler."""
-        super().__init__(Entropy(axis=axis), DeterministicSampler())
+        super().__init__(Entropy(axis=axis), TopKSampler())
