+++ pyrelational/strategies/classification/ratio_confidence_strategy.py
@@ -1,7 +1,7 @@
 """Active learning using ratio based confidence uncertainty measure."""
 
+from pyrelational.batch_mode_samplers import TopKSampler
 from pyrelational.informativeness import RatioConfidence
-from pyrelational.samplers import DeterministicSampler
 from pyrelational.strategies.classification.classification_strategy import (
     ClassificationStrategy,
 )
@@ -15,4 +15,4 @@ class RatioConfidenceStrategy(ClassificationStrategy):
 
     def __init__(self, axis: int = -1):
         """Initialize the strategy with the ratio confidence scorer and a deterministic scorer for classification."""
-        super().__init__(RatioConfidence(axis=axis), DeterministicSampler())
+        super().__init__(RatioConfidence(axis=axis), TopKSampler())
