+++ pyrelational/strategies/regression/greedy_strategy.py
@@ -1,7 +1,7 @@
 """Greedy Strategy Module."""
 
+from pyrelational.batch_mode_samplers import TopKSampler
 from pyrelational.informativeness import AverageScorer
-from pyrelational.samplers.samplers import DeterministicSampler
 from pyrelational.strategies.regression.regression_strategy import RegressionStrategy
 
 
@@ -14,4 +14,4 @@ class GreedyStrategy(RegressionStrategy):
 
     def __init__(self, axis: int = 0):
         """Initialize the strategy with the mean prediction scorer and a deterministic scorer for regression."""
-        super().__init__(AverageScorer(axis=axis), DeterministicSampler())
+        super().__init__(AverageScorer(axis=axis), TopKSampler())
